fix: 헬스장선택, 기본수업시간 설정 스크롤 이슈 수정 및 파일분리
