+++ src/app/class-time-setting/page.tsx
@@ -1,11 +1,7 @@
 import { ClassTimeSettingPage } from '@/page/schedule';
 
 const ClassTimeSetting = () => {
-  return (
-    <div>
-      <ClassTimeSettingPage />
-    </div>
-  );
+  return <ClassTimeSettingPage />;
 };
 
 export default ClassTimeSetting;
